Add support for prompt caching (#30)
* Add cache control to `TextContent`

Add cache control to `TextContent` and validate that the value gets
serialied correctly for the Converse API as well as for the invoke model
API with Claude and Nova LLMs.

* Nova: use `cachePoint` instead of `cache_point` in JSON payload

* Support Bedrock Converse cache read/write tokens in usage

* Validate that Claude works

* Add docstring to `_handle_content` for Converse API

Add a docstring explaining why the function returns a list of content
rather than a single content object.

* Make library's API mirror the Converse API

After discussing, we determined that we preferred the Bedrock Converse's
API style of treating a cache point as its own content type rather than
the Nova/Claude invoke model-style API of making the caching part of
each individual content type.

Update the API the library exposes to mirror the Bedrock style, and
update the transformation for the 3 APIs we use to support the new API.

* Add a base class for `ClaudeCacheableContent`

Reduce boilerplate declaring the caching fields in the Claude content
types by introducing this base type.

* Add `NovaCacheableContent`

This additional type reduces boilerplate and matches the type pattern
that exists with the Claude API.

* Claude `should_request_prompt_cache` -> `should_cache`

Simpler name. More alignment with Nova API.

* Do not use computed fields in Claude/Nova models

As pointed out in code review, these models are intended to be transit
types. They should prefer making it easy to reason about the JSON they
generate rather than making it ergonomic to instantiate for a developer.

* Fix docstring