Make JSON mode in Claude models more robust (#31)
* make json_mode in Claude models more robust

Problem: Previously, we enforced JSON mode in Claude models by setting "{" as the response prefix. This ensured that the LLM generated JSON but it did not ensure that it stopped generating once the JSON was finished, so it would sometimes output extra text after the JSON which broke the JSON parsing.

Solution: Change the response prefix to "<json>\n{" and set "</json>" as a stop sequence.

* fix env var name in test script